# Hardware/Actuator/NomadBLDC/Tools/GUI/nomad-bldc-app/openapi.json
{"openapi":"3.0.2","info":{"title":"NomadBLDC API","description":"API for Nomad BLDC motor controllers","contact":{"name":"Quincy Jones","url":"http://www.implementedrobotics.com","email":"quincy@implementedrobotics.com"},"license":{"name":"MIT","url":"https://opensource.org/licenses/MIT"},"version":"0.1"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/devices":{"get":{"summary":"Get Devices","operationId":"get_devices_devices_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"summary":"Create Device","operationId":"create_device_devices_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddNomadBLDCModel"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/devices/{device_id}":{"get":{"summary":"Get Device","operationId":"get_device_devices__device_id__get","parameters":[{"required":true,"schema":{"title":"Device Id","type":"string"},"name":"device_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/devices/connect/{device_id}":{"get":{"summary":"Connect Device","operationId":"connect_device_devices_connect__device_id__get","parameters":[{"required":true,"schema":{"title":"Device Id","type":"string"},"name":"device_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/devices/disconnect/{device_id}":{"get":{"summary":"Disconnect Device","operationId":"disconnect_device_devices_disconnect__device_id__get","parameters":[{"required":true,"schema":{"title":"Device Id","type":"string"},"name":"device_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/devices/status/{device_id}":{"get":{"summary":"Connect Device","operationId":"connect_device_devices_status__device_id__get","parameters":[{"required":true,"schema":{"title":"Device Id","type":"string"},"name":"device_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/stream":{"post":{"summary":"Stream Telemetry","operationId":"stream_telemetry_stream_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TelemetryData"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AddNomadBLDCModel":{"title":"AddNomadBLDCModel","required":["name","can_bus_id"],"type":"object","properties":{"name":{"title":"Name","type":"string","description":"Friendly name of the Nomad BLDC Device"},"can_bus_id":{"title":"Can Bus Id","type":"integer","description":"The CAN BUS Identifier for Nomad BLDC Device"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"TelemetryData":{"title":"TelemetryData","required":["nomad_bldc_id","registers"],"type":"object","properties":{"nomad_bldc_id":{"title":"Nomad Bldc Id","type":"string","description":"The UID of the Nomad BLDC Device"},"registers":{"title":"Registers","type":"array","items":{"type":"integer"},"description":"List of registers to stream telemetry data"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}